requirements.txt

instaloader
matplotlib
mtcnn
tensorflow
Pillow
Keras
